Bugfix for Period Group user permission in Employee

+++ workwise/employee_201/doctype/employee/employee.py
@@ -108,8 +108,9 @@ class Employee(Document):
 				frappe.permissions.add_user_permission("Department", parent_department, self.user_id)
 		if self.period_group and self.user_id:
 			user_perm = frappe.db.sql(""" SELECT `for_value` FROM `tabUser Permission` WHERE `user` = %s AND `allow` = "Period Group" """,(self.user_id) , as_dict=1)
-			if user_perm != self.period_group:
+			if user_perm and user_perm[0].for_value != self.period_group:
 				frappe.db.sql("""DELETE FROM `tabUser Permission` WHERE `user` = %s AND allow = "Period Group" """,(self.user_id),as_dict=True)
+			if not user_perm:
 				frappe.permissions.add_user_permission("Period Group", self.period_group, self.user_id)
 		if not self.period_group:
 			frappe.db.sql("""DELETE FROM `tabUser Permission` WHERE `user` = %s AND allow = "Period Group" """,(self.user_id),as_dict=True)
